+++ Ratz/Assets/Scripts/Beat Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a697f81b38660a44fb5b6d0401ddafbf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
